Add files via upload
The best way to use the code is to draw a rough sketch of the molecule that you want to draw, then select the backbone and branches accordingly. And don't forget to name the locations! Hopefully in the future I'll add more features (such as free electrons & ionic charges, resonance structure, wavey bonds to symbolize a long bond, etc.) and make the code more user friendly.